# doc/api.yaml
openapi: 3.0.0
info:
  title: WasaPhoto - OpenAPI 3.0
  description: | 
    This is the api specification for WasaPhoto, an educational purpose based project, 
    to keep in touch with your friends by sharing photos of special moments.  
    - [The github private project repository](https://github.com/luigidannibale/Wasa)
    
    Functional design specifications: 
    
    Each user will be presented with a stream of photos (images) in reverse chronological order, with
    information about when each photo was uploaded (date and time) and how many likes and comments
    it has. The stream is composed by photos from “following” (other users that the user follows). Users
    can place (and later remove) a “like” to photos from other users. Also, users can add comments to any
    image (even those uploaded by themself). Only authors can remove their comments.

    Users can ban other users. If user Alice bans user Eve, Eve won’t be able to see any information about
    Alice. Alice can decide to remove the ban at any moment.
    
    Users will have their profiles. The personal profile page for the user shows: the user’s photos (in reverse
    chronological order), how many photos have been uploaded, and the user’s followers and following.
    Users can change their usernames, upload photos, remove photos, and follow/unfollow other users.
    Removal of an image will also remove likes and comments.
    
    A user can search other user profiles via username.
    
...

...
  contact:
...
security:
  - UserAuth: []
paths:        
...
    /users:
      post:
        operationId: doLogin
        security: []
        tags: 
          - User
          
        summary: Logs in a user
        description: |-
          If the user does not exist, it will be created,
          and an identifier is returned.
          If the user exists, the user identifier is returned.        
        requestBody:
          description: Username of the user          
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User/properties/username"
                  
        responses:
          "200":
            description: User log-in action successful
            content:
              application/json:
                schema:
                    $ref: "#/components/schemas/Message"
          "201":
            description: New user log-in action successful
            content:
              application/json:
                schema:
                    $ref: "#/components/schemas/Message"
          "400": 
            $ref : "#/components/responses/BadRequest" 
          "500":
            $ref : "#/components/responses/InternalServerError"        
      get:
        operationId: getUserProfile
        tags: 
          - User
          
        summary: Returns the profile of the user.
        description: |
          The profile of the user is returned, so all the details about the user.
        parameters:
          - name: username
            in: query
            description: The username of the user to find 
            required: true
            schema:
              $ref : "#/components/schemas/User/properties/username"
        responses: 
          "200":
            description: Successful, returning the user
            content:
              application/json:      
                schema:
                  $ref : "#/components/schemas/Message"
          "400":
            $ref : "#/components/responses/BadRequest" 
          "401":
            $ref : "#/components/responses/Unauthorized" 
          "404":
            $ref : "#/components/responses/NotFound" 
          "500":
            $ref : "#/components/responses/InternalServerError"      
    
    #updateUser     - LastMod : 11/12 
...
    /users/{userID}/profile/username:      
      parameters:
      - $ref : "#/components/parameters/userID"
      put:
        operationId: setMyUserName
        tags:
          - User
          
        summary: Changes the username        
        description: |
          Takes a user (by the id) and a name and sets the name as the new username of such user.           
        requestBody:
          description: The username to set
          content:
            application/json:
              schema:
                $ref : "#/components/schemas/User/properties/username"          
        responses:
          "200":
            $ref : "#/components/responses/OK"
          "201":
            $ref : "#/components/responses/Created"
          "400":
            $ref : "#/components/responses/BadRequest"
          "401":
            $ref : "#/components/responses/Unauthorized"                    
          "403":
            $ref : "#/components/responses/Forbidden"
          "404":
            $ref : "#/components/responses/NotFound"
          "409":
            $ref : "#/components/responses/Conflict"
          "500":
            $ref : "#/components/responses/InternalServerError"   
        
    #getMyStream  - LastMod : 30/11 
...
components:  
  securitySchemes: # LastMod - 2/12
    UserAuth: 
      type: http
      scheme: bearer
  parameters: # LastMod - 28/11
    userID: 
      name: userID
      in: path
      description: Id of the user
      required: true
      schema:
        $ref : "#/components/schemas/User/properties/id"
...
  schemas:  #LastMod - 11/12
...
    Message:       
      description: a simple text message
      type: string
      example: "This is the outcome"
      pattern: ".{0,1024}"
      minLength: 0
      maxLength: 1024
...
  responses: # LastMod - 2/12
    OK: 
      description: The request has been completed successfully
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"
    Created: 
      description: The request has been completed successfully, resulting in a creation
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"
    BadRequest: 
      description: The request is bad asked
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"
    Unauthorized: 
      description: This request is unauthorized
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"    
    Forbidden: 
      description: Resource(s) not found
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"
    NotFound: 
      description: Resource(s) not found
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"    
    Conflict: 
      description: The request creates a conflict
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"
    InternalServerError: 
      description: Internal server error has occurred
      content: 
        application/json:
          schema:
            $ref : "#/components/schemas/Message"